+++ .idea/workspace.xml
@@ -2,9 +2,17 @@
 <project version="4">
   <component name="ChangeListManager">
     <list default="true" id="869db707-2ddf-4e85-b700-72708b6fc648" name="Default Changelist" comment="">
+      <change afterPath="$PROJECT_DIR$/depricated_code/gnarfle_00_bst_tools.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/garfle_module_02_parse_for_job_titles.py" beforeDir="false" afterPath="$PROJECT_DIR$/garfle_module_02_parse_for_job_titles.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/garfle_module_02_parse_for_job_titles.py" beforeDir="false" afterPath="$PROJECT_DIR$/depricated_code/garfle_module_02_parse_for_job_titles.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/gnarfle_00_bst_tools.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/gnarfle_01_from_indeed_to_aws.py" beforeDir="false" afterPath="$PROJECT_DIR$/depricated_code/gnarfle_01_from_indeed_to_aws.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/gnarfle_a0_commands.py" beforeDir="false" afterPath="$PROJECT_DIR$/gnarfle_a0_commands.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/gnarfle_indeed.py" beforeDir="false" afterPath="$PROJECT_DIR$/gnarfle_indeed.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/gnarfle_one_shots.py" beforeDir="false" afterPath="$PROJECT_DIR$/gnarfle_one_shots.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/login_info.txt" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/read_the_indeed_search_queue.py" beforeDir="false" afterPath="$PROJECT_DIR$/depricated_code/read_the_indeed_search_queue.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/whack_words_test.py" beforeDir="false" afterPath="$PROJECT_DIR$/whack_words_test.py" afterDir="false" />
     </list>
     <option name="EXCLUDED_CONVERTED_TO_IGNORED" value="true" />
     <option name="SHOW_DIALOG" value="false" />
@@ -31,13 +39,13 @@
     <property name="settings.editor.selected.configurable" value="com.jetbrains.python.configuration.PyActiveSdkModuleConfigurable" />
   </component>
   <component name="RecentsManager">
-    <key name="CopyFile.RECENT_KEYS">
-      <recent name="E:\Software\gnarfle" />
-    </key>
     <key name="MoveFile.RECENT_KEYS">
       <recent name="E:\Software\gnarfle\depricated_code" />
       <recent name="E:\Software\gnarfle" />
       <recent name="E:\Software\gnarfle\SQL" />
+    </key>
+    <key name="CopyFile.RECENT_KEYS">
+      <recent name="E:\Software\gnarfle" />
     </key>
   </component>
   <component name="RunDashboard">
@@ -65,7 +73,7 @@
       <option name="IS_MODULE_SDK" value="true" />
       <option name="ADD_CONTENT_ROOTS" value="true" />
       <option name="ADD_SOURCE_ROOTS" value="true" />
-      <option name="SCRIPT_NAME" value="$PROJECT_DIR$/garfle_module_02_parse_for_job_titles.py" />
+      <option name="SCRIPT_NAME" value="E:\Software\gnarfle\depricated_code\garfle_module_02_parse_for_job_titles.py" />
       <option name="PARAMETERS" value="" />
       <option name="SHOW_COMMAND_LINE" value="false" />
       <option name="EMULATE_TERMINAL" value="false" />
@@ -86,7 +94,7 @@
       <option name="IS_MODULE_SDK" value="true" />
       <option name="ADD_CONTENT_ROOTS" value="true" />
       <option name="ADD_SOURCE_ROOTS" value="true" />
-      <option name="SCRIPT_NAME" value="$PROJECT_DIR$/gnarfle_01_from_indeed_to_aws.py" />
+      <option name="SCRIPT_NAME" value="E:\Software\gnarfle\depricated_code\gnarfle_01_from_indeed_to_aws.py" />
       <option name="PARAMETERS" value="" />
       <option name="SHOW_COMMAND_LINE" value="false" />
       <option name="EMULATE_TERMINAL" value="false" />
@@ -243,7 +251,7 @@
     <breakpoint-manager>
       <breakpoints>
         <line-breakpoint enabled="true" type="python-line">
-          <url>file://$PROJECT_DIR$/garfle_module_02_parse_for_job_titles.py</url>
+          <url>file://$PROJECT_DIR$/depricated_code/garfle_module_02_parse_for_job_titles.py</url>
           <option name="timeStamp" value="2" />
         </line-breakpoint>
       </breakpoints>
