+++ gnarfle_a0_commands.py
@@ -1,6 +1,6 @@
 
 
-from gnarfle_01_from_indeed_to_aws import  build_on_the_indeed_search_list
+from depricated_code.gnarfle_01_from_indeed_to_aws import  build_on_the_indeed_search_list
 
 
 build_on_the_indeed_search_list()
