--- login_info.txt
@@ -1,5 +1,0 @@
-# master username
-
-db_string:gnarfle.cv3nmvm2rhep.us-east-2.rds.amazonaws.com
-user_id:Beldar
-password:CyZn7rLx1qAplQWd
